New module render/renderSignin.j to allow res.render('signin') call with proper context.
