+++ routes/signin.js
@@ -4,26 +4,9 @@
 
 var express = require('express')
 var oidc = require('../oidc')
-var settings = require('../boot/settings')
-var mailer = require('../boot/mailer')
 var authenticator = require('../lib/authenticator')
 var passwordless = require('./passwordless')
-var qs = require('qs')
-var providers = require('../providers')
-
-var providerInfo = {}
-var providerNames = Object.keys(providers)
-for (var i = 0; i < providerNames.length; i++) {
-  providerInfo[providerNames[i]] = providers[providerNames[i]]
-}
-var visibleProviders = {}
-// Only render providers that are not marked as hidden
-Object.keys(settings.providers).forEach(function (providerID) {
-  if (!settings.providers[providerID].hidden) {
-    visibleProviders[providerID] = settings.providers[providerID]
-  }
-})
-
+var renderSignin = require('../render/renderSignin')
 /**
  * Signin Endpoint
  */
@@ -38,13 +21,7 @@ module.exports = function (server) {
     oidc.verifyClient,
     oidc.validateAuthorizationParams,
     function (req, res, next) {
-      res.render('signin', {
-        params: qs.stringify(req.query),
-        request: req.query,
-        providers: visibleProviders,
-        providerInfo: providerInfo,
-        mailSupport: !!(mailer.transport)
-      })
+      renderSignin(res, req.query)
     })
 
   /**
@@ -73,21 +50,11 @@ module.exports = function (server) {
     function (req, res, next) {
       authenticator.dispatch(req.body.provider, req, res, next, function (err, user, info) {
         if (err) {
-          res.render('signin', {
-            params: qs.stringify(req.body),
-            request: req.body,
-            providers: visibleProviders,
-            providerInfo: providerInfo,
-            mailSupport: !!(mailer.transport),
+          renderSignin(res, req.body, {
             error: err.message
           })
         } else if (!user) {
-          res.render('signin', {
-            params: qs.stringify(req.body),
-            request: req.body,
-            providers: visibleProviders,
-            providerInfo: providerInfo,
-            mailSupport: !!(mailer.transport),
+          renderSignin(res, req.body, {
             formError: info.message
           })
         } else {
